fundacionmichis/requirements.txt
django-pwa
social-auth-app-django
